⛸️🌕 ↝ [SSM-2]: Moving unused items and missions to archived api routes
