+++ .vercelignore
@@ -1,3 +1,5 @@
 .env.local
 .env
 citizen/buckets
+citizen/supabase
+citizen/archive
